rename asset.type.ts to types.ts
