+++ web/src/components/molecules/Common/CommentsPanel/Comment.tsx
@@ -12,7 +12,7 @@ import Input from "@reearth-cms/components/atoms/Input";
 import Tooltip from "@reearth-cms/components/atoms/Tooltip";
 import UserAvatar from "@reearth-cms/components/atoms/UserAvatar";
 import { User } from "@reearth-cms/components/molecules/AccountSettings/types";
-import { Comment } from "@reearth-cms/components/molecules/Asset/asset.type";
+import { Comment } from "@reearth-cms/components/molecules/Asset/types";
 import { dateTimeFormat } from "@reearth-cms/utils/format";
 
 const { TextArea } = Input;
