+++ web/src/components/molecules/Common/CommentsPanel/Thread.tsx
@@ -1,5 +1,5 @@
 import { User } from "@reearth-cms/components/molecules/AccountSettings/types";
-import { Comment } from "@reearth-cms/components/molecules/Asset/asset.type";
+import { Comment } from "@reearth-cms/components/molecules/Asset/types";
 
 import { CommentList } from "./CommentList";
