+++ web/src/components/molecules/Common/Form/AssetItem/hooks.ts
@@ -1,8 +1,8 @@
 import { useState, useCallback } from "react";
 
 import { UploadFile } from "@reearth-cms/components/atoms/Upload";
-import { Asset } from "@reearth-cms/components/molecules/Asset/asset.type";
 import { UploadType } from "@reearth-cms/components/molecules/Asset/AssetList";
+import { Asset } from "@reearth-cms/components/molecules/Asset/types";
 import { useProject, useWorkspace } from "@reearth-cms/state";
 
 export default (
