+++ web/src/components/molecules/Common/Form/GroupItem/index.tsx
@@ -4,8 +4,8 @@ import { useCallback, useMemo, MouseEvent } from "react";
 import Collapse from "@reearth-cms/components/atoms/Collapse";
 import Icon from "@reearth-cms/components/atoms/Icon";
 import { UploadFile } from "@reearth-cms/components/atoms/Upload";
-import { Asset } from "@reearth-cms/components/molecules/Asset/asset.type";
 import { UploadType } from "@reearth-cms/components/molecules/Asset/AssetList";
+import { Asset } from "@reearth-cms/components/molecules/Asset/types";
 import {
   AssetField,
   ReferenceField,
