+++ web/src/components/molecules/Content/Form/fields/ComplexFieldComponents/AssetField.tsx
@@ -1,7 +1,7 @@
 import Form from "@reearth-cms/components/atoms/Form";
 import { UploadFile } from "@reearth-cms/components/atoms/Upload";
-import { Asset } from "@reearth-cms/components/molecules/Asset/asset.type";
 import { UploadType } from "@reearth-cms/components/molecules/Asset/AssetList";
+import { Asset } from "@reearth-cms/components/molecules/Asset/types";
 import AssetItem from "@reearth-cms/components/molecules/Common/Form/AssetItem";
 import MultiValueAsset from "@reearth-cms/components/molecules/Common/MultiValueField/MultiValueAsset";
 import FieldTitle from "@reearth-cms/components/molecules/Content/Form/FieldTitle";
