+++ web/src/components/molecules/Content/Form/fields/ComplexFieldComponents/GroupField.tsx
@@ -1,7 +1,7 @@
 import Form, { FormInstance } from "@reearth-cms/components/atoms/Form";
 import { UploadFile } from "@reearth-cms/components/atoms/Upload";
-import { Asset } from "@reearth-cms/components/molecules/Asset/asset.type";
 import { UploadType } from "@reearth-cms/components/molecules/Asset/AssetList";
+import { Asset } from "@reearth-cms/components/molecules/Asset/types";
 import GroupItem from "@reearth-cms/components/molecules/Common/Form/GroupItem";
 import MultiValueGroup from "@reearth-cms/components/molecules/Common/MultiValueField/MultiValueGroup";
 import FieldTitle from "@reearth-cms/components/molecules/Content/Form/FieldTitle";
