+++ web/src/components/molecules/Request/Details/CommentList.tsx
@@ -2,7 +2,7 @@ import styled from "@emotion/styled";
 
 import List from "@reearth-cms/components/atoms/List";
 import { User } from "@reearth-cms/components/molecules/AccountSettings/types";
-import { Comment } from "@reearth-cms/components/molecules/Asset/asset.type";
+import { Comment } from "@reearth-cms/components/molecules/Asset/types";
 import ThreadCommentMolecule from "@reearth-cms/components/molecules/Request/Details/Comment";
 
 type Props = {
