+++ web/src/components/molecules/Schema/FieldModal/FieldDefaultInputs/index.tsx
@@ -1,6 +1,6 @@
 import { UploadFile } from "@reearth-cms/components/atoms/Upload";
-import { Asset } from "@reearth-cms/components/molecules/Asset/asset.type";
 import { UploadType } from "@reearth-cms/components/molecules/Asset/AssetList";
+import { Asset } from "@reearth-cms/components/molecules/Asset/types";
 import {
   AssetSortType,
   SortDirection,
